split node into headless/javafx (closes #88)
