+++ eclair-node-javafx/src/main/deploy/package/windows/Eclair (Alpha 3).ico
