+++ eclair-node-javafx/src/main/resources/gui/commons/globals.css
