+++ eclair-node-javafx/src/main/resources/gui/commons/images/close.png
