+++ eclair-node-javafx/src/main/resources/gui/commons/images/connection_icon.png
