+++ eclair-node-javafx/src/main/resources/gui/commons/images/eclair-fit.png
