+++ eclair-node-javafx/src/main/resources/gui/commons/images/eclair-shape.png
