+++ eclair-node-javafx/src/main/resources/gui/commons/images/in-out-11.png
