+++ eclair-node-javafx/src/main/resources/gui/commons/images/success_icon.png
