+++ eclair-node-javafx/src/main/resources/gui/commons/images/warning_icon.png
