+++ eclair-node-javafx/src/main/resources/gui/main/channelPane.fxml
