+++ eclair-node-javafx/src/main/resources/gui/main/main.fxml
