+++ eclair-node-javafx/src/main/resources/gui/main/notificationPane.fxml
