+++ eclair-node-javafx/src/main/resources/gui/main/notifications.fxml
