+++ eclair-node-javafx/src/main/resources/gui/modals/about.fxml
