+++ eclair-node-javafx/src/main/resources/gui/modals/openChannel.fxml
