+++ eclair-node-javafx/src/main/resources/gui/modals/receivePayment.fxml
