+++ eclair-node-javafx/src/main/resources/gui/modals/sendPayment.fxml
