+++ eclair-node-javafx/src/main/resources/gui/splash/splash.fxml
