+++ eclair-node-javafx/src/main/scala/fr/acinq/eclair/gui/controllers/NotificationPaneController.scala
