+++ eclair-node-javafx/src/main/scala/fr/acinq/eclair/gui/stages/AboutStage.scala
