+++ eclair-node-javafx/src/main/scala/fr/acinq/eclair/gui/utils/GUIValidators.scala
