+++ eclair-node-javafx/src/test/resources/logback-test.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration scan="true" debug="false">
+
+    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+        <target>System.out</target>
+        <encoder>
+            <pattern>%date{HH:mm:ss.SSS} %highlight(%-5level) %X{akkaSource} - %msg%ex{12}%n</pattern>
+        </encoder>
+    </appender>
+
+    <!--appender name="CONSOLEWARN" class="ch.qos.logback.core.ConsoleAppender">
+        <target>System.out</target>
+        <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
+            <level>WARN</level>
+        </filter>
+        <encoder>
+            <pattern>%-5level %X{akkaSource} - %msg%ex{12}%n</pattern>
+        </encoder>
+    </appender-->
+
+    <!--appender name="FILE" class="ch.qos.logback.core.FileAppender">
+        <file>eclair.log</file>
+        <append>false</append>
+        <encoder>
+            <pattern>%-5level %X{akkaSource} - %msg%ex{12}%n</pattern>
+        </encoder>
+    </appender-->
+
+    <logger name="fr.acinq.eclair.Pipe" level="DEBUG" />
+    <logger name="fr.acinq.eclair.crypto.TransportHandler" level="DEBUG" />
+
+    <root level="INFO">
+        <!--appender-ref ref="FILE"/>
+        <appender-ref ref="CONSOLEWARN"/-->
+        <appender-ref ref="CONSOLE"/>
+    </root>
+
+</configuration>
