+++ pom.xml
@@ -10,6 +10,7 @@
     <modules>
         <module>eclair-core</module>
         <module>eclair-node</module>
+        <module>eclair-node-javafx</module>
     </modules>
 
     <description>A scala implementation of the Lightning Network</description>
